requirements.txt
flask
python-dotenv
google-generativeai
requests
beautifulsoup4
psycopg2-binary